Dodany spring, Tesy JMeter, Program testujący Selenium

+++ lab1/lab01/src/main/java/com/tmejs/db/service/NauczycielManager.java
@@ -33,7 +33,7 @@ public interface NauczycielManager {
 
     public int addNauczyciel(Nauczyciel person);
     
-    public Nauczyciel getNauczyciel(Long id);
+    public Nauczyciel getNauczyciel(Integer id);
     
     public Boolean updateNauczyciel(Nauczyciel nauczyciel);
